backend/requirements.txt
asgiref==3.9.1
Brotli==1.1.0
cffi==1.17.1
cssselect2==0.8.0
Django==5.2.4
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
Faker==37.4.2
fonttools==4.59.0
Markdown==3.8.2
pillow==11.3.0
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
pyphen==0.17.2
python-dotenv==1.1.1
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2025.2
validate_docbr==1.11.1
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1
